Rearranging certain parts of code structure
