+++ app/scripts/general-setup.js
@@ -1,6 +1,6 @@
 import { QuizzerDataClass } from "../modules/AppData/AppData.js";
 import UI_InterfaceClass from "../modules/UI/UI_Interface.js";
-import { save_UI_Entries } from "../modules/util/HandlersHelpers.js";
+import { HandlerHelpersClass } from "../modules/util/Helpers.js";
 import { CategoriesURL } from "../modules/util/URL.js";
 import API_ServiceClass from "../Services/API_Service.js";
 import { LocalDataPersistenceClass } from "../Services/PersistentService.js";
@@ -13,6 +13,7 @@ const quizzerData = new QuizzerDataClass();
 const selectElement = UI_Interface.getElements('#language-options')[0];
 const localDataPersistenceService = new LocalDataPersistenceClass();
 const Router = new RouterService();
+const HandlerHelpers = new HandlerHelpersClass()
 
 quizzerData.setData(['Quiz Categories', await API_Service.fetchData(CategoriesURL).then(data => data.trivia_categories)])
 quizzerData.getData('Quiz Categories').forEach(categoryObj => {
@@ -26,10 +27,10 @@ quizzerData.getData('Quiz Categories').forEach(categoryObj => {
 function save_UI_Config_Entries(event) {
     const selectedCategoryElement = UI_Interface.getElements(`#language-options option:nth-of-type(${selectElement.selectedIndex + 1})`)[0];
 
-    save_UI_Entries(quizzerData.setConfigData, 'selectedCategoryName', selectedCategoryElement.innerText);
-    save_UI_Entries(quizzerData.setConfigData, 'selectedCategoryId', selectedCategoryElement.value);
+    HandlerHelpers.saveEntries(quizzerData.setConfigData, 'selectedCategoryName', selectedCategoryElement.innerText);
+    HandlerHelpers.saveEntries(quizzerData.setConfigData, 'selectedCategoryId', selectedCategoryElement.value);
 
-    save_UI_Entries(quizzerData.setConfigData, 'selectedDifficulty', Array.from(UI_Interface.getElements("input[type = 'radio']")).find(radioElement => radioElement.checked)?.value);
+    HandlerHelpers.saveEntries(quizzerData.setConfigData, 'selectedDifficulty', Array.from(UI_Interface.getElements("input[type = 'radio']")).find(radioElement => radioElement.checked)?.value);
 
     localDataPersistenceService.saveData('Quizzer Config Data', Object.fromEntries(quizzerData.getConfigData().entries()));
     Router.goToRoute(UI_Interface.getAttribute([this], 'href')[0])
